// src/enums/google-recaptcha-enterprise-reason.ts
export enum GoogleRecaptchaEnterpriseReason {
    InvalidReasonUnspecified = 'INVALID_REASON_UNSPECIFIED',
    UnknownInvalidReason = 'UNKNOWN_INVALID_REASON',
    Malformed = 'MALFORMED',
    Expired = 'EXPIRED',
    Dupe = 'DUPE',
    SiteMismatch = 'SITE_MISMATCH',
    Missing = 'MISSING',
    BrowserError = 'BROWSER_ERROR',
}